requirements.txt
python-telegram-bot==21.6
aiohttp==3.10.10
aiofiles==24.1.0
requests==2.32.3
python-dotenv==1.0.1
yookassa==3.8.0
